backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfminer.six
python-docx
sentence-transformers
scikit-learn
spacy